fix link query error bugs
