+++ pages/index.js
@@ -201,10 +201,17 @@ const Index = () => {
                   </div>
                   {!isEmpty(name) && !isEmpty(password) ? (
                     <Link
+                      href={{
+                        pathname: '/sender/main',
+                        query: { name: name, num: password },
+                      }}
+                    >
+                      {/*                   
+                    <Link
                       href={`/sender/[main]`}
                       // p/[receiver]?name=${name}&pw=${password}&id=${id}
                       as={`/sender/main?name=${name}&num=${password}`}
-                    >
+                    > */}
                       <Buttons
                         content="생성하기"
                         full={true}
